% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamMixRHLP.R
\docType{class}
\name{ParamMixRHLP-class}
\alias{ParamMixRHLP-class}
\alias{ParamMixRHLP}
\title{A Reference Class which contains parameters of a MixRHLP model.}
\arguments{
\item{G}{The number of clusters.}
}
\description{
ParamMixRHLP contains all the parameters of a MixRHLP model.
}
\section{Fields}{

\describe{
\item{\code{fData}}{\link{FData} object representing the sample.}

\item{\code{K}}{The number of regimes (mixture components).}

\item{\code{p}}{The order of the polynomial regression.}

\item{\code{q}}{The dimension of the logistic regression. For the purpose of
segmentation, it must be set to 1.}

\item{\code{variance_type}}{Numeric indicating if the model is homoskedastic
(\code{variance_type} = 1) or heteroskedastic (\code{variance_type} = 2).}

\item{\code{Wg}}{Parameters of the logistic process.
\eqn{Wg = w_{g1},\dots,w_{gK-1}}{Wg = (wg1,\dots,wgK-1)} is a matrix of dimension
\eqn{(q + 1, K - 1)}, with \emph{q} the order of the logistic regression.}

\item{\code{betag}}{Parameters of the polynomial regressions.
\eqn{\betag = (\beta_{g1},\dots,\beta_{gK})}{\beta = (\betag1,\dots,\betagK)} is
a matrix of dimension \eqn{(p + 1, K)}, with \emph{p} the order of the
polynomial regression.}

\item{\code{sigma2_g}}{The variances for the \emph{K} regimes. If MixRHLP model is
homoskedastic (\emph{variance_type} = 1) then sigma2_g is a matrix of size
\eqn{(1, 1)}, else if MixRHLP model is heteroskedastic then sigma2_g is a matrix
of size \eqn{(K, 1)}.}

\item{\code{pi_jgk}}{Logistic proportions for cluster of size \eqn{(mn, K, G)}.}

\item{\code{alpha_g}}{Cluster weights of size \eqn{(G, 1)}.}
}}

\section{Methods}{

\describe{
\item{\code{init_hlp(try_algo)}}{  initialize the Hidden Logistic Process}

\item{\code{initRegressionParam(Xg, g, try_algo)}}{  Initialize the Regresssion model with Hidden Logistic Process}
}}

\seealso{
\link{FData}
}
