% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatMixRHLP.R
\docType{class}
\name{StatMixRHLP-class}
\alias{StatMixRHLP-class}
\alias{StatMixRHLP}
\title{A Reference Class which contains statistics of a MixRHLP model.}
\description{
StatMRHLP contains all the parameters of a \link[=ParamMixRHLP]{MixRHLP} model.
}
\section{Fields}{

\describe{
\item{\code{h_ig}}{Fuzzy segmentation matrix of size \eqn{(n, G)} representing the
post probabilities prob(curve|cluster_g)}

\item{\code{c_ig}}{Hard partition logical matrix of dimension \eqn{(n, G)}
obtained by the Maximum a posteriori (MAP) rule:
\eqn{c_{ig} = 1 \ \textrm{if} \ c_{ig} = \textrm{arg} \ \textrm{max}_{k} \
P(c_i = g | Y, W, \beta);\ 0 \ \textrm{otherwise}}{c_ig = 1 if c_ig =
arg max_k P(c_i = g | Y, W, \beta); 0 otherwise}, \eqn{k = 1,\dots,G}.}

\item{\code{klas}}{Column matrix of the labels issued from \code{z_ik}. Its elements are
\eqn{klas(i) = k}, \eqn{k = 1,\dots,K}.}

\item{\code{Ex_g}}{Column vector of dimension m for each g \emph{(m,G)}. \code{Ex_g} is the curve expectation
: sum of the polynomial components \eqn{\beta_{gk} \times X_{i}}{\betak x X_i}
weighted by the logistic probabilities \code{pij_gk}:
\eqn{Ex_g(j) = \sum_{k = 1}^{K} pi_jik \times \beta_{gk} \times X_{i}}{Ey(i) =
\sum_{k=1}^K pi_jik x \betak x X_i}, \eqn{i = 1,\dots,n}.}

\item{\code{log\_lik}}{Numeric. Log-likelihood of the MixRHLP model.}

\item{\code{com_loglik}}{Numeric. Complete log-likelihood of the MixRHLP model.}

\item{\code{stored_loglik}}{List. Stored values of the log-likelihood at each EM
iteration.}

\item{\code{stored_com_loglik}}{List. Stored values of the complete log-likelihood at each EM iteration.}

\item{\code{tau_ijgk}}{Matrix of dimension \emph{(nm,K)} for each g (g=1,...,G), representing the fuzzy segmentation
for the cluster g prob(y_{ij}|kth_segment,cluster_g)}

\item{\code{log_tau_ijgk}}{the logarithmic values for tau_ijgk}

\item{\code{BIC}}{Numeric. Value of the BIC (Bayesian Information Criterion)
criterion. The formula is \eqn{BIC = log\_lik - nu \times
\textrm{log}(n) / 2}{BIC = log\_lik - nu x log(n) / 2} with \emph{nu} the
degree of freedom of the MixRHLP model.}

\item{\code{ICL}}{Numeric. Value of the ICL (Integrated Completed Likelihood)
criterion. The formula is \eqn{ICL = com\_loglik - nu \times
\textrm{log}(n) / 2}{ICL = com_loglik - nu x log(n) / 2} with \emph{nu} the
degree of freedom of the MixRHLP model.}

\item{\code{AIC}}{Numeric. Value of the AIC (Akaike Information Criterion)
criterion. The formula is \eqn{AIC = log\_lik - nu}{AIC = log\_lik - nu}.}

\item{\code{cpu_time}}{Numeric. Average executon time of a EM step.}

\item{\code{log_fg_xij}}{Loglikelihood for the n_g curves of cluster g. Matrix of dimension \eqn{(n, G)}}

\item{\code{log_alphag_fg_xij}}{Matrix of dimension \eqn{(n, G)}}

\item{\code{polynomials}}{Matrix of size \eqn{(m, K, G)} giving the values of
\eqn{\beta_{k} \times X_{i}}{\betak x X_i}, \eqn{i = 1,\dots,m}.}

\item{\code{weighted_polynomials}}{Matrix of size \eqn{(m, K, G)}}
}}

\section{Methods}{

\describe{
\item{\code{MAP()}}{   calculate a partition by applying the Maximum A Posteriori Bayes
   allocation rule}
}}

\seealso{
\link{ParamMixRHLP}, \link{FData}
}
