% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMixRHLP.R
\docType{class}
\name{ModelMixRHLP-class}
\alias{ModelMixRHLP-class}
\alias{ModelMixRHLP}
\title{A Reference Class which represents a fitted MixRHLP model.}
\description{
ModelMixRHLP represents a \link[=ModelMixRHLP]{MixRHLP} model for which parameters have
been estimated.
}
\section{Fields}{

\describe{
\item{\code{paramMixRHLP}}{A \link{ParamMixRHLP} object. It contains the estimated values of the parameters.}

\item{\code{statMixRHLP}}{A \link{StatMixRHLP} object. It contains all the statistics associated to the MixRHLP model.}
}}


\seealso{
\link{ParamMixRHLP}, \link{StatMixRHLP}
}
